Rewrite: Simplify logic, add more helper functions.
Trying to control all the checking and toggling of the whitelist from
one function is too complicated. Now the functions have been separated
out, with each clearly showing what it does.

Checking functions now return three states instead of two, which
allows a sort of hierarchy of logic. First you check for the specific
url, if that is undefined, then check the stem. This way the specific
url has precedence, but a false does not override the stem result.

Cleaning urls without slashes is fixed by adding slashes if necessary
before and after cleaning.