Rewrite: Integrate improvements from common.js.
Move message sending after the whitelist is updated. content.js actually
checks the whitelist instead of a simple toggle and if the whitelist is
checked before it is updated the result is sometimes wrong. This makes
it slower, but always correct.

Integrate checkWhitelist change from common.js. checkWhitelist now
returns a boolean specifying if dark mode is true(on) or false(off) and
the result must be interpreted by the functions that set the window
state.

Add functions for setting the state of the toggle stem button.