Manifest: Remove "allFrames" option from content scripts...
In some cases, mostly when an iframe is loaded after a page is loaded,
an iframe is inverted when the parent page is inverted. This causes
problems because the chrome extension api does not inject css into all
iframes.

The solution for now is to only have dark mode invert the parent page
and essentially ignore iframes.