Feature: Add "Ctrl-Shift-I" Toggle Invert feature.
This was not as simple as it seems.

The Chrome extension api is very picky about where some methods are
called. The api commands here communicate with the background javascript
"background.js" which then communicates with the front end "content.js".

This required a restructure due to lack of foresight. Previously the
logic for adding or removing a url to the whitelist lived in "popup.js".
This was moved to "common.js" where it could be used in both the popup
and be called from "content.js" for keyboard shortcut toggling.