Feature: Implement add dark mode to iframes.
This implementation requires refactoring the way the current page
receives a message to turn on/off dark mode.

Due to XSS concerns executing a "content" script inside an iframes
violates Chrome's (or really any browser's) cross domain javascript
rules. Meaning that the previous implementation where the content js
communicated with the background js cannot work inside iframes.

To get around this the chrome api offers a "chrome.tabs.executeScript"
function. With the "all_frames" option this executes js on every iframe.
But the caveat is this needs to be sent (executed) from the background.
So this function piggybacks on the context menu events and sends the
relevant script every time a context menu event, essentially a window or
tab change, is fired.

Iframes are marked with the html attribute "dark-mode-iframe=true" and
styled differently, avoiding double inversion, to match the parent page.
The iframe styling, for some reason, is really fragile and easy to
break. More investigation is needed into why this happens.