Bug Fix: Popup communicates with messages.
Previously the popup directly read and parsed the whitelist. Now when
the popup opens or a button is clicked, a message is sent to the
background page which responds with an object containing the current
status of dark-mode.

This object is grabbed by a listener and parsed to setup/change the
popup html.

This is much more reliable than the previous implementation. Just try
clicking the toggle button a bunch of times!