Structure: Add `PersistentStorage` class.
`PersistentStorage` is the parent class for working with `chrome.storage` objects.

`PersistentStorage` handles common functions for working with objects
that need to be kept both readily accessible (not behind a callback),
frequently accessed, and saved to `chrome.storage`.

`PersistentStorage` will be used as the parent class for the forthcoming
url whitelist class and the global settings class, a feature that is
coming in the future.