Structure: Standardize setting names and types.
Add static `SettingId` class that handles every Id for all settings.

This will be used when a setting name needs to be specified with a
string.

This addition keeps all setting names in location. It ensures that the
names used to identify things are the same across multiple classes and
files.

This is not a DRY solution on purpose. Having autocomplete for these
values is more important than a clever solution. And if anything needs
to be added or changed it can be done, its just a little harder.