CSS: Render content css with postcss.
To implement css with multiple combinations of dark mode settings
requires a more efficient css generator than either less or sass offers.

Postcss with postcss-js syntax allows for writing css in javascript that
is then compiled to css. This is most necessary for handling multiple
contrast values from the `data-dark-mode-contrast` attribute.

The build process is done with gulp. I did not figure out a way to
render the css with the gulp-postcss module, so rendering is done by
including the required modules in `dark-mode.js` and executing the file
with node.

The less build process was also incorperated in the gulpfile.

The css is now built with the running the gulp default task with `gulp`.