Refactor: Improve popup ui responsiveness.
Have popup run in a headless state:

The popup initially recieves values from the background, then jQuery
renders value changes which are sent to the background. No effort is
made to confirm that values are correct in the popup if they are
altered.

This is more efficient because waiting for the listener after a value is
changes is unpredictable and slow, and because most UI elements are
custom jQuery things and require a initilization every time the state
changes.