Feature: Use webNavigation api to track frameIds.
To properly invert an iframe the inversion of the base frame (parent) matters.
For some reason `-webkit-filter: invert()` crosses the iframe boundary. This
makes sense for an iframe that is embedded one "level" within the parent page.
But an iframe that is embedded multiple "levels" is inverted incorrectly.

This is solved by using the inversion at each level to calculate the inversion
of the next level. This means knowing the state (Dark & Hue) of each frame,
   derived from the url.