CSS: Add dark support for `twitterwidget` elements in the shadow dom
This was tricky to figure out but essentially a shadow dom element is an
element whose style is not affected by the page style unless
specifically targeted. It is NOT an iframe and can be targeted directly
with CSS using the `::shadow` selector. Right now this only targets
`twitterwidget` elements but ideally we could use a glob and select
every shadow dom element to apply dark styles