Popup: Introduce Collapse component
`Collapse` wraps a component with a clickable collapse label.

One caveat: material-ui's Collapse component uses `overflow: hidden` to
wrap the collapse component, meaning nothing will render outside the
bounds of the component. Practically, this meant we could not use a
popup to display tooltips because we cannot guarantee they will render
within the bounds of the parent collapse component.