Refactor: Watch for shadow dom events with `MutationObserver`
For a reason which I don't know/understand, events in the shadow dom don't
broadcast to the parent page. Instead we are using `MutationObserver` to
react to `childList` and `subTree` changes, inspired by
https://stackoverflow.com/questions/28230040/web-component-how-to-listen-to-shadow-dom-load-event?rq=1.

This makes `invertInsideShadowDom` run much more frequently, but because this
function is lightweight, the user should see no regression in
performance or features.